- move typing stuff to _types - move the tests to new test_types file accordingly - add type check to `fillna`
